refactor: 프로필 수정 페이지 리펙토링
